+++ exts/events/events.py
@@ -478,9 +478,10 @@ class EventHandler(commands.Cog, CogMixin):
 
         if before.embeds:
             data = before.embeds[0]
-            if data.type == "image" and not self.is_url_spoiler(
-                before.content, data.url
-            ):
+            if data.type == "image":
+                # and not self.is_url_spoiler(
+                #     before.content, data.url
+                # ):
                 e.set_image(url=data.url)
 
         if before.attachments:
@@ -503,7 +504,7 @@ class EventHandler(commands.Cog, CogMixin):
                     inline=False,
                 )
 
-        return await logCh.send(embed=e)
+        return await logCh.send(content=before.channel.mention, embed=e)  # type: ignore
 
     @commands.Cog.listener("on_message_delete")
     async def onMessageDelete(
@@ -538,7 +539,9 @@ class EventHandler(commands.Cog, CogMixin):
             else message.content
         )
 
-        return await logCh.send(embed=e)
+        return await logCh.send(
+            content=message.channel.mention, embed=e  # type: ignore
+        )
 
     @commands.Cog.listener("on_member_update")
     async def onMemberUpdate(